Refactor url params handling
[#162085672]
